Tools: Added task for upload of wrapper api docs.
The task can be used for both ios and android as they both have the same structure on S3.

Command: `gulp upload-wrapper-apidocs --release-version <x.y.z> --bucket <your-bucket>`. Or for only uploading specific wrapper(s): `gulp upload-wrapper-apidocs --release-version <x.y.z> --bucket <your-bucket> --only-upload-wrappers ios,android`

There is a slight chance that the user can mistype the version. Ideally we should retrieve the version number from the built api docs files. But I cannot see that it is possible at the current moment for either of the wrappers.

Also did a simple refactoring (moved code) from the uploadS3.js file to a more specific task as some functions were specific to that task.
